# docs/openapi.json
{
  "components": {
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Error"
      }
    },
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "RS256 JWT, or an API key of the form `key.\u003cuuid\u003e.\u003c64-hex-secret\u003e`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for the OCI-Janus registry management BFF (registry-management), generated from the service route table.\n\n**Authentication.** Send an `Authorization: Bearer \u003ctoken\u003e` header — either an RS256 JWT (browser/dashboard) or an API key `key.\u003cuuid\u003e.\u003c64-hex-secret\u003e` (CI/automation). Operations marked with a lock require authentication; API-key principals carry no roles, so admin-only operations return 403.\n\n**Conventions.** Pagination is cursor-based (`page_token` + `per_page`); errors return `{\"error\": \"...\"}`. This document describes paths, methods, path parameters, and the auth requirement exactly; request/query and response body schemas are being enriched incrementally. The identity/session API (login, SSO, MFA, API keys) is served separately by registry-auth and is not included here.",
    "title": "OCI-Janus Management API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/orgs/{org}/members": {
      "get": {
        "operationId": "get_api_v1_orgs_org_members",
        "parameters": [
          {
            "in": "path",
            "name": "org",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "GET /api/v1/orgs/{org}/members",
        "tags": [
          "orgs"
        ]
      },
      "post": {
        "operationId": "post_api_v1_orgs_org_members",
        "parameters": [
          {
            "in": "path",
            "name": "org",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "POST /api/v1/orgs/{org}/members",
        "tags": [
          "orgs"
        ]
      }
    },
    "/api/v1/orgs/{org}/members/{assignmentID}": {
      "delete": {
        "operationId": "delete_api_v1_orgs_org_members_assignmentID",
        "parameters": [
          {
            "in": "path",
            "name": "org",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "assignmentID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "DELETE /api/v1/orgs/{org}/members/{assignmentID}",
        "tags": [
          "orgs"
        ]
      }
    },
...
    "/api/v1/repositories/{org}/{repo}/members": {
      "get": {
        "operationId": "get_api_v1_repositories_org_repo_members",
        "parameters": [
          {
            "in": "path",
            "name": "org",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "GET /api/v1/repositories/{org}/{repo}/members",
        "tags": [
          "repositories"
        ]
      },
      "post": {
        "operationId": "post_api_v1_repositories_org_repo_members",
        "parameters": [
          {
            "in": "path",
            "name": "org",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "POST /api/v1/repositories/{org}/{repo}/members",
        "tags": [
          "repositories"
        ]
      }
    },
    "/api/v1/repositories/{org}/{repo}/members/{assignmentID}": {
      "delete": {
        "operationId": "delete_api_v1_repositories_org_repo_members_assignmentID",
        "parameters": [
          {
            "in": "path",
            "name": "org",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "assignmentID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "DELETE /api/v1/repositories/{org}/{repo}/members/{assignmentID}",
        "tags": [
          "repositories"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same origin as the gateway",
      "url": "/"
    }
  ],
...
}